requirements.txt
azure-functions
azure-kusto-data
openai
matplotlib
pandas
fpdf
requests
azure-communication-email
azure-identity
azure-core

scikit-learn
seaborn
numpy
statsmodels
